many: remove unnecessary formatting in prints, logs and errors (#14363 backport)
With the go1.23 release `go vet` has started to complain about calls to
functions with a formatting string and varargs that use a non-constant
string and don't have anything to format. So this change replaces all
the unnecessary calls to Printf and its cousins with the non-formatting
equivalent. Also adds those functions where we didn't have them before
(e.g., logger/).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>
Co-authored-by: Miguel Pires <miguel.pires@canonical.com>