i/p/requestprompts: add timeout for prompts
When there has been no retrieval of prompt details for a given user
after a particular duration, expire all outstanding prompts for that
user, as this suggests that there is no prompt client running for that
user.

If a user retrieves prompt details or interacts with prompts in some
way, such as by retrieving all prompts, a prompt by ID, or attempting to
reply to a prompt, bump the timeout to a much longer timeout, since this
indicates that a prompt client is running for that user.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>