o/snapstate: do not keep the state lock while removing snap files
Do not take the state lock when removing snap files, since it is not known how
long the removal will take. We have observed a system in the field, where there
were issues with I/O on the storage device and unlink() took significant time,
this preventing snapd from functioning normally as the state lock was held.

Related: SNAPDENG-26617

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>