o/i/apparmorprompting: record notices in goroutines
During snapd startup, the state lock is held while the prompting-related
backends are created. However, the backends may try to record notices
if, e.g., some rules on disk have expired before they are read by load.

Similarly, during snapd shutdown, the state lock may be held while the
interfaces requests manager is being stopped.

If recording a notice happens synchronously during such situations, it
results in a deadlock on the state lock. Thus, the notices must be
recorded asynchronously.

Note: this means that the order of notices is no longer guaranteed in
the prompt and rule backends, but this is okay, since clients should
always be prepared to consult the actual prompting endpoints as a single
source of truth, and notice data is just a convenient shortcut if it is
known that a prompt or rule has already been removed.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>