o/snapstate,kernel: allow drivers in $SNAP_DATA (#14346)
* o/snapstate,kernel: allow drivers in $SNAP_DATA

Consider modules and firmware in $SNAP_DATA when creating the drivers
tree. Drivers in that place can be created by hooks from
kernel-modules components. The drivers are expected in
$SNAP_DATA/{modules,firmware}.

* kernel: do not log as notice no fw/modules in component file

Now these can be dynamically created so logging the siutation as debug
should be enough.

* kernel: add dynamic-modules field to kernel.yaml

dynamic-modules will contain a folder where subfolder
{modules,firmware} might be present. For the moment only $SNAP_DATA or
${SNAP_DATA} are allowed.

* o/snapstate: consider dynamic modules only if specified in kernel.yaml