o/i/apparmorprompting: record notices in goroutines (#14405)
* o/i/apparmorprompting: record notices in goroutines

During snapd startup, the state lock is held while the prompting-related
backends are created. However, the backends may try to record notices
if, e.g., some rules on disk have expired before they are read by load.

Similarly, during snapd shutdown, the state lock may be held while the
interfaces requests manager is being stopped.

If recording a notice happens synchronously during such situations, it
results in a deadlock on the state lock. Thus, the notices must be
recorded asynchronously.

Note: this means that the order of notices is no longer guaranteed in
the prompt and rule backends, but this is okay, since clients should
always be prepared to consult the actual prompting endpoints as a single
source of truth, and notice data is just a convenient shortcut if it is
known that a prompt or rule has already been removed.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/i/apparmorprompting: increase wait notices timeout in tests

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/i/apparmorprompting: do not manually set notice timestamp

There is a fix in state/notices.go to ensure that no two notices can
have the same timestamp by increasing the timestamp of any duplicates by
1ns, but that only applies to notices with an unset Time in the
AddNoticeOptions. Thus, with manually set times, we can get identical
timestamps. For now, stop manually setting timestamps.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/i/apparmorprompting: fix bug in tests while waiting for no notices

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* tests: add test of snapd startup with apparmor prompting enabled

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>