o/ifacestate,tests: apparmor prompting not running if manager creation failed
If creation of the interfaces requests manager fails for some reason, we
need to mark the manager's internal `useAppArmorPrompting` value to
false so that calls to `AppArmorPromptingRunning()` will return false,
and we don't try to access a nil-pointer manager backend.

Extend the apparmor-prompting-snapd-startup test to cover this scenario.

Additionally, fix a few syntax inconsistencies in the spread test while
we're at it.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>