o/snapstate: create shared snap directory during link-snap
Fix a problem with parallel installs and handling of the shared snap directory,
eg /snap/foo, for snap instances /snap/foo_bar. A bug was identified which was a
race between a removal of the shared directory in the discard-snap handler of
one instance, and creating of the same shared directory in setup-snap handler of
another instances if both changes happened to be executing in parallel. The
existing conflict resolution does not consider the instances of the same snap to
be in conflict, since with exception of the shared snap directory and/or
aliases, there should be no collision.

In detail, the shared directory was created in setup-snap handler, which runs
before link-snap, and so before the instance-being-installed is effectively
added to the state. At the same time, discard-snap handler would check the state
for other snap instances, but since the new installation isn't there yet, there
would be none, and so the shared directory would get removed.

Related: SNAPDENG-31335

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>