snapdtool: enable FIPS on classic without snap
Enable Go FIPS runtime when executing on a classic distro when system is in FIPS
mode. This allows us to run a FIPS variant of the snapd deb package on a FIPS
enabled system and ensure that snapd will now attempt to setup the FIPS
execution mode. Note, this is transparent for snapd binaries that have not been
built with FIPS Go toolhchain.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>