cmd: rename SNAP_MOUNT_DIR to STATIC_SNAP_MOUNT_DIR
Given that most tools now detect the snap mount directory dynamically, picking
either /var/lib/snapd/snap or /snap, depending on what is actually being used,
it is the right time to rename the remaining option to the more appropriate
name of static snap mount directory. The name differs because the value is
baked into the files shipped with the distribution-specific packaging. This
includes systemd units, environment generators and the management script.

We still need this value and it is not going away anytime soon, but it is not
the only value and so I think that separating SNAP_MOUNT_DIR from
STATIC_SNAP_MOUNT_DIR is well, apt.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>