o/registrystate: create modify-registry tasks
Adds a helper that attaches the appropriate tasks to gather approval,
save, commit and notify of a change to a registry. The specific approval
gathering mechanism is described in SD133. This doesn't include the
task/hook handlers that deal with the specific but the whole picture can
be seen in https://github.com/canonical/snapd/pull/14283

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>