s/apparmor: expand apparmor prompting support checks
Prompting cannot be supported if the kernel notification socket does not
exist. Thus, add a check for its presence to PromptingSupportedByFeatures.

However, this alone is not sufficient. The notification socket may be
used for other purposes as well. When this becomes the case, a directory
will be added at `/sys/kernel/security/apparmor/features/policy/notify`.
If that directory exists, and it contains a file called `user`, which
contains the list of prompting mediation classes, then prompting is
supported on the system. If no file called `user` exists in that
directory, but the directory itself exists, then we know prompting is
not supported.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>