Revert "tests/main/interfaces-snap-interfaces-requests-control: disable on debian-sid"
This reverts commit 66eb8624ce6620fcce738983881d350ebf24cb1f.

Debian Sid reports prompting support in its AppArmor kernel and parser
features, the former because the upstream kernel has extended
permissions merged (but non-functional), and the latter because we use
vendored AppArmor with snapd. But the kernel does not actually support
AppArmor prompting, as this requires patches from the Ubuntu kernel
which have not yet been upstreamed.

Now that we check for the presence of the kernel notification socket as
well as the AppArmor kernel and parser features, we can correctly
identify that prompting is not supported on Debian Sid.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>