audio-playback: Add support for snapped pipewire
A snapped pipewire requires being able to create the pulse
folder inside the global XDG_RUNTIME_DIR folder (not the snap
own folder). Also, it requires to be able to create both the
pipewire-[0-9]-manager and its lock file.

This patch adds that capability to the audio-playback slot;
thus, a snap offering that slot will be able to create the
corresponding global sockets.