interfaces/builtin/custom-device: allow to explicitly identify matching device in udev tagging block
There is an implicit assumption that the `kernel` entry in `udev-tagging`
snippet matches the device path listed in devices. While this is often true for
most devices, there are some where the kernel device is exposed under a
completely different path under /dev, specifically msr[0-9], are exposed at
/dev/cpu/0-9/msr, rather than /dev/msr[0-9]. Hence the implicit rules made it
impossible to describe a device in a way which would result in the right access
rules.

Introduces a new attribute `for-device` in in the udev-tagging snippets which
allows to establish a match between said snippet and an actual device listed in
devices section.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>