tests/core/remodel-base: fix the test when store base has different revision (#14452)
On 2024-08-29 a new revision of core18 was pushed latest/edge to the store,
which caused the test to revert back to a revision different than in the seed
which is older. In this scenario, the restore snippet would incorrectly identify
no need to a reboot and so the test would break. Snapd when forced to
reinitialize the system would issue reboot, since the installed core18 is
different than one in the seed, but the test would not trigger a reboot through
spread.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>