cmd/snap-update-ns: re-create layouts referencing content
A layout entry constructed from the same location that is used as a
target of a content target, is susceptible to propagation issues.  When
the content is connected, the mounted content is propagated to from the
content target location to layout target location. When the content is
disconnected the unmount event does not propagate, as that causes
propagation issues described by umount(2), quoted in full below:

   umount() and shared mounts
	Shared mounts cause any mount activity on a mount, including
	umount() operations, to be forwarded to every shared mount in
	the peer group and every slave mount of that peer group. This
	means that umount() of any peer in a set of shared mounts will
	cause all of its peers to be unmounted and all of their slaves
	to be unmounted as well.

	This propagation of unmount activity can be particularly
	surprising on systems where every mount is shared by default.
	On such systems, recursively bind mounting the root directory of
	the filesystem onto a subdirectory and then later unmounting
	that subdirectory with MNT_DETACH will cause every mount in the
	mount namespace to be lazily unmounted.

	To ensure umount() does not propagate in this fashion, the mount
	may be remounted using a mount(2) call with a mount_flags
	argument that includes both MS_REC and MS_PRIVATE prior to
	umount() being called.

For this reason, snap-update-ns has been using MNT_DETACH preceded by
MS_PRIVATE for many years now.

The consequence of this asymmetry is that content exposed through a
layout stays visible even though through the layout, even after being
disconnected.

Out of several possible attempts to fix this, the simplest one seems to
be not to reuse any layout entries that overlap content connections that
are not going to be reused. A layout entry is then re-constructed from
the directory where the now-disconnected content used to be attached at,
behaving exactly as one would expect.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-31645

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>