client,daemon: added dedicated error kinds for former validation errors
Rather than keep the "validation" errors as special cases of the invalid
field error kind, instead add dedicated error kinds for them.

These errors are:
- reply pattern does not match requested path
- reply permissions do not include all requested permission
- a rule with conflicting path pattern and permissions already exists

For the latter in partifular, it's not really an "invalid field" error,
since the combination of the "path-pattern" and "permissions" fields
results in a conflict with an existing rule, rather than one single
field being the problem.

There was additional confusion because for all these errors, the
received values are well-formed, but the reply is rejected due to the
state of the system.

With these now having dedicated error kinds, the json error values can
be much simpler, including just the requested path and replied pattern,
the requested and replied permissions, and the details of the rule
conflicts, respectively.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>