tests: make prompting-related tests work on any supported ubuntu system
Ubuntu systems older than 24.04 are capable of supporting AppArmor
prompting if they are running a new enough Ubuntu kernel.

In particular, the Ubuntu 22.04 image on GCP runs a 6.5 kernel which
supports prompting. Thus, if run on GCP, Jammy supports prompting, but
if run on lxd, it does not.

Therefore, checks for Ubuntu version are insufficient for spread to
identify whether prompting should be supported. Instead, check for the
presence of the `prompt` directive in the permissions listed in
`/sys/kernel/security/apparmor/features/policy/permstable32`.

Any Ubuntu system which has `prompt` in that file should support
prompting. This can't be the entirety of the check, because those
permissions have been upstreamed to the Linux kernel, but other pieces
which are required for prompting to function have not yet been
upstreamed. Thus, Debian Sid reports supporting the prompt directive,
even though prompting is not supported.

Therefore, we must check that the system is Ubuntu, that it's not core,
and that the `prompt` directive is supported.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>