many: consolidate prompting errors and add prompting error kinds (#14438)
* many: consolidate prompting errors and add prompting error kinds

Rather than each prompting-related package defining and using their own
error types, define all named errors in the `interfaces/prompting`
package, and use them from elsewhere. This requires some adjustments to
error messages.

Additionally, add corresponding error kinds for all errors which can
occur from API requests and are expected to be handled in a particular
way by clients. These include most of the errors defined in the
prompting package, but omits some which are expected to be strictly
internal errors.

Thus, clients can rely on error kinds rather than attempting to parse
error messages from the API.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* client,daemon: treat prompting backends being closed as internal errors and remove error kinds

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon,interfaces/prompting: move RuleConflictError into i/prompting

In doing so, only implement marshalling for the RuleConflictError type
in the daemon.

Additionally, when a rule conflict is encountered for one permission,
continue testing remaining permissions, so that all conflicts can be
reported in the error.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon: remove unused prompting errorIsOneOf helper

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: simplify prompting error kinds and improve error robustness

Rather than having an error kind for each prompting-related error,
instead reduce to four total error kinds for prompting:
- ErrorKindAppArmorPromptingNotRunning
- ErrorKindInterfacesRequestsPromptNotFound
- ErrorKindInterfacesRequestsRuleNotFound
- ErrorKindInterfacesRequestsInvalidField

Errors of kind ErrorKindInterfacesRequestsInvalidField make use of the
API error's `Value` field. Namely, the `Value` field is a map from field
name to a new `invalidFieldValue` struct. This struct contains an
`"invalid-reason"` field and a `"metadata"` field.

Invalid field errors can occur for three categories of reasons:
1. unsupported value when there's a fixed set of supported values
2. parse error when a field value is malformed
3. validation error when a field value is valid, but the state of the
   system causes there to be an error (e.g. conflicting rule)

For the first two categories, we define a common error struct each which
can be adapted to particular error values within a given category. To
avoid a circular import from the interfaces/prompting/patterns package,
these error struct definitions are placed in
interfaces/prompting/errortypes, while all but the one error required by
patterns defined in interfaces/prompting/errors.go as constructors of
those error structs.

For the third category, three particular errors can occur, and each has
a dedicated error struct which includes error-specific details in
queryable fields, as well as an identifier of which error occurred. This
provides prompting clients (and their users) information required to
adjust and retry their request such that it is valid with the current
state of the system.

As part of introducing these new error structs, some prompting-related
error messages have changed. In particular, some details which were
previously included in error message strings are now included in struct
fields which are marshalled in the API error value.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: move prompting errors to interfaces/prompting/errors package

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting/errors: rename error constructors from err... to ...Error

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* client,daemon: added dedicated error kinds for former validation errors

Rather than keep the "validation" errors as special cases of the invalid
field error kind, instead add dedicated error kinds for them.

These errors are:
- reply pattern does not match requested path
- reply permissions do not include all requested permission
- a rule with conflicting path pattern and permissions already exists

For the latter in partifular, it's not really an "invalid field" error,
since the combination of the "path-pattern" and "permissions" fields
results in a conflict with an existing rule, rather than one single
field being the problem.

There was additional confusion because for all these errors, the
received values are well-formed, but the reply is rejected due to the
state of the system.

With these now having dedicated error kinds, the json error values can
be much simpler, including just the requested path and replied pattern,
the requested and replied permissions, and the details of the rule
conflicts, respectively.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* interfaces/prompting: simplify errors, use helper functions for creating new error instances

Attempt to simplify prompting errors. Drop values that were not really error
values and replace them with function helpers for creating corresponding error
values.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* daemon: rename and minor refactor to handling of prompting errors

Drop additional struct types in favor of simple auxiliary type definitions where
possible.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* daemon,i/p/errors: adjusted prompting error internals and organization

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon: flatten metadata field from prompting errors

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* many: consolidate error kinds for reply not matching request

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon: remove "invalid-" prefix from prompting invalid field error json

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon: add tests for prompting api error responses

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon,i/prompting: make unsupported value errors always produce list of values

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* i/prompting/errors: use strutil.Quoted

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>
Co-authored-by: Maciej Borzecki <maciej.borzecki@canonical.com>