tests: new approach to run tests in ubuntu and debian using snapd deb from the repo
The idea of this change is to change de default behaviour for the deb
pacage used in our spread tests. To simulate a more real scenario, and
having snapd snap build from local, this change is adding as default a
new scenario to have snapd installed from the deb repository.

Then the deb package is saved in the same place as it was built in
during the test, making easy to have both scenarios working together.

It is a requirement also so be able to run the tests using a built snapd
deb as we are currently doing.