s/apparmor: expand apparmor prompting support checks (#14448)
* s/apparmor: expand apparmor prompting support checks

Prompting cannot be supported if the kernel notification socket does not
exist. Thus, add a check for its presence to PromptingSupportedByFeatures.

However, this alone is not sufficient. The notification socket may be
used for other purposes as well. When this becomes the case, a directory
will be added at `/sys/kernel/security/apparmor/features/policy/notify`.
If that directory exists, and it contains a file called `user`, which
contains the list of prompting mediation classes, then prompting is
supported on the system. If no file called `user` exists in that
directory, but the directory itself exists, then we know prompting is
not supported.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: add missing feature to mocked features list in prompting support test

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* features: remove superfluous test of apparmor function which has in-situ test

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* Revert "tests/main/interfaces-snap-interfaces-requests-control: disable on debian-sid"

This reverts commit 66eb8624ce6620fcce738983881d350ebf24cb1f.

Debian Sid reports prompting support in its AppArmor kernel and parser
features, the former because the upstream kernel has extended
permissions merged (but non-functional), and the latter because we use
vendored AppArmor with snapd. But the kernel does not actually support
AppArmor prompting, as this requires patches from the Ubuntu kernel
which have not yet been upstreamed.

Now that we check for the presence of the kernel notification socket as
well as the AppArmor kernel and parser features, we can correctly
identify that prompting is not supported on Debian Sid.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* Revert "tests/main/apparmor-prompting-snapd-startup: account for vendored apparmor on 22.04"

This reverts commit 1ec413226a455a90be6faca5bfc1e5cae7ed1ad2.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* Revert "tests: account for prompting supported on >= 22.04 with internal apparmor"

This reverts commit a4e46bbc710731e1c327f2bab938d4dca056d3e7.

On Ubuntu LTS releases < 24.04, prompting is still not supported by the
standard (non-hwe) kernels. This is reflected in tests, now that
AppArmor prompting support checks test for the presence of the kernel
notification socket.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor,tests: improve comments around apparmor prompting

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* tests: add debug checks for prompting support

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* tests: make prompting-related tests work on any supported ubuntu system

Ubuntu systems older than 24.04 are capable of supporting AppArmor
prompting if they are running a new enough Ubuntu kernel.

In particular, the Ubuntu 22.04 image on GCP runs a 6.5 kernel which
supports prompting. Thus, if run on GCP, Jammy supports prompting, but
if run on lxd, it does not.

Therefore, checks for Ubuntu version are insufficient for spread to
identify whether prompting should be supported. Instead, check for the
presence of the `prompt` directive in the permissions listed in
`/sys/kernel/security/apparmor/features/policy/permstable32`.

Any Ubuntu system which has `prompt` in that file should support
prompting. This can't be the entirety of the check, because those
permissions have been upstreamed to the Linux kernel, but other pieces
which are required for prompting to function have not yet been
upstreamed. Thus, Debian Sid reports supporting the prompt directive,
even though prompting is not supported.

Therefore, we must check that the system is Ubuntu, that it's not core,
and that the `prompt` directive is supported.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: add test case when policy/notify/user file exists but does't include file

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: adjust prompting unsupported reason error messages

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: prompting support checks for permstable32_version >= 2

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: add TODO about using dirs instead of rootPath

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/c/configcore,s/apparmor: mock aa root path  in prompting tests

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: trim whitespace from permstable32_version before parsing

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* s/apparmor: adjust prompting unsupported reason error messages to reference prompting

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>