i/p/patterns: clean path pattern variants before parsing
Path patterns may contain `/./` or `/../` constructions, which we should
handle similarly to how they should be in standard paths. That is:

- /foo/./bar is equal to /foo/bar
- /foo/../bar is equal to /bar

This allows us to treat identical path patterns (e.g. /foo/./bar and
/foo/bar) as identical, rather than two distinct patterns.

This is useful for convenience so that the client may, for example,
when requested /foo/bar/file.txt, reply with /foo/bar/file.txt/../* to
grant access to all files in the parent directory of the requested path.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>