tests: detect loop devices with deleted files as an invariant
When a loop device is created with a given backing file, the file itself
may be deleted. The kernel keeps the reference to the file alive but it
also adjusts the name of the file to have the suffix " (deleted)".

We should not have loop devices with this property across tests.
They might indicate that a mount sticks around and leaks through a
reference somehow, or that the low-level loop-back utilities are buggy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>