o/registrystate: add registry tx task handlers and helpers
Adds a RegistryManager which implements handlers for tasks related to
the committing of registry transactions. These do things like committing
the transaction internally, clean up state once that's done or on error,
etc. This doesn't include logic to deal with concurrent writers, that
will be added later once the registrystate side of this is added as
well.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>