tests: udpate the github workflow to run tests suggested by spread-filter tool
Through this change we will run the tests suggested by the spread-filter
tool following the rules located in tests/lib/spread/rules which defined
for each system.

For nested tests, this ruleset will make sure that:
. Always that the nested.sh lib or the assertions dir are updated, all
the nested test are executed
 . When an specific nested test is updated, then this is executed
. If the "Run Nested" label is added to the PR, all the nested tests
will be executed as well.
 . In master branch, all nested test are executed
 . In other scenarios, nested tests are not executed.

Also, in master branch, all the tests are executed (the tests defined in
the matrix for each system).