tests: show that snapctl can be used with bare base
The bare base snap has no liraries and no dynamic linker. Show that even
in such environment it is possible to use the snapctl program.

This test also helps in testing upcoming changes to the snapd-specific
dynamic linker, namely in the fact that snapd linker is not affecting
snapctl.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-32203

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>