interfaces: generalize constant folding for patterns
The existing constant folding mechanism for sequences-of-literals is now generalized
to inject the literal into the following item, possibly reducing the complexity
of the tree at the cost of making the tree contain longer strings.

The tests are not fully adjusted, as the render test measures precise indices that
were too tedious to fix on Friday.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>