daemon,i/p/requestprompts,o/i/apparmorprompting: explicitly indicate client activity
Callers of the requestprompts methods can specify whether they indicate
that a prompting client is present and active, and thus the prompt
expiration timeout should be reset.

Previously, any call to these methods was implicitly treated as
activity. Now, these functions have the ability to be called internally
or while testing or debugging in a way which does not cause the timeout
to be reset.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>