cmd/snap-bootstrap: XXX: run ps aux once a second
We are seeing oom killing snap-bootstrap that takes over a 1.6GB of
memory on a 2GB system used in CI. We don't have a theory as to what is
going on yet. This patch is not mergable but might help us see what is
is happening on the path that leads to failure.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>