cmd/snap-update-ns: allow detaches to propagate
When snap-update-ns chooses not to reuse a mount entry it attempts to unmount
it by detaching the mount entry. Before that is done the mount entry used to be
switch to private propagation, so that the unmount operation was isolated.

Disable this logic so that unmounts (and detaches) propagate normally. The
mount namespace of the snap is mostly switched to slaved propagation, with the
sole exception of /media or /run/media, which uses bidirectional propagation by
design.

This patch is experimental and is not ready for being merged yet.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-31645

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>