tests: show that snap from snapd deb can be bundled
Developers may want to bundle the snap executable from the snapd project
into their application to get access to the snap APIs in a convenient
form.

This test shows that this can be safely done for as long as the snapd
deb is used as the source (not snapd snap) and that the ABI of snapd.deb
and the application snap's base match.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-32203

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>