tests: generalize cifs-home to have an nfs variant
This test is much more modern and easier to reason about. It is also somewhat different
in that it doesn't really exericse all the various configurations where nfs/cifs is detected,
that is handled with unit tests (and we know the detection part is relatively weak).

Instead the test focuses that a given user may be "remote home" and that the special
permissions work as expected.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>