cmd/snap-confine: pass uid/gid to initial tmpfs
We mount the tmpfs which will be used as the rootfs of the mount
namespace of the snap without any options. Since at the time were are
running as root (user) but keep the original group of the calling user,
this captures the gid of the user in the mount options.

Pass uid=0,gid=0 explicitly, so that it is not an issue.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-33380

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>