tests/lib/fakestore/store: support auto-refreshes in fakestore
Fakestore was not implemented with auto-refreshes in mind.
When a refresh action request was received any unknown snap-ids
caused a 400 error and on test systems it is expected to have some
snaps available not directly being tested.

Fakestore logic was updated to ignore those unknown snap-ids if the
incoming refresh action was an auto-refresh which could be detected
by inspecting the "Snap-Refresh-Reason" header and checking that it
is set to "scheduled".

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>