tests: replace nfs support test with generalized cifs test (#14562)
* tests: add okfail executable

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: generalize cifs-home to have an nfs variant

This test is much more modern and easier to reason about. It is also somewhat different
in that it doesn't really exericse all the various configurations where nfs/cifs is detected,
that is handled with unit tests (and we know the detection part is relatively weak).

Instead the test focuses that a given user may be "remote home" and that the special
permissions work as expected.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: remove the nfs-support test

This test is now replaced by generalized cifs-support test. Not as many systems are covered
but the test is much more robust and easier to work with.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* spread: use set -xeu

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: remove explicit package removal

Packages are automatically removed by a higher-level restore logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>